Cleaned up Data-management
