+++ bin/startup-exit.js
@@ -1,0 +1,25 @@
+import fs from "fs";
+import {newVideos, videoList} from "./dataFiles.js";
+import {addToQueue, tryDownload} from "./downloadManager.js";
+
+let cleanUpAndExit = function () {
+
+    fs.writeFileSync('./data/videoData.json', JSON.stringify(videoList));
+    fs.writeFileSync('./data/newVideos.json', JSON.stringify(newVideos));
+    process.exit(0);
+};
+
+let restoreProgress = function () {
+
+    for(let video of videoList) {
+
+        if(!video.downloaded) {
+
+            addToQueue(video);
+        }
+    }
+
+    tryDownload().catch();
+};
+
+export {restoreProgress, cleanUpAndExit}
